add itemControl for window view
