+++ SimplyTotalCommander/MainWindow.xaml
@@ -6,113 +6,8 @@
         xmlns:local="clr-namespace:SimplyTotalCommander"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
-    <Grid>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition />
-        </Grid.ColumnDefinitions>
-        <TabControl Grid.ColumnSpan="3" Margin="0,0,-0.4,0">
-            <TabItem Header="File List"
-                     ToolTip="Most common file details">
-                <WrapPanel Orientation="Horizontal" Margin="0,0,-2,-2">
-                    <!-- *** Showing table with details of files from list. *** -->
-                    <DataGrid x:Name="dataGridOfFiles" 
-                              IsReadOnly="True"
-                              AutoGenerateColumns="True" 
-                              SelectionUnit="CellOrRowHeader"
-                              Margin="0,0,399.6,0"
-                              Width="623"
-                              VerticalAlignment="Top" 
-                              Height="275"
-                              SelectionChanged="DataGridOfFiles_SelectionChanged"
-                  >
-                        <DataGrid.FrozenColumnCount>1</DataGrid.FrozenColumnCount>
-                    </DataGrid>
-                    <!-- *** Select or search for a file in currnet directory. *** -->
-                    <ComboBox SelectionChanged="Button_Click" 
-                              x:Name="SelectedFile" 
-                              IsEditable="True" 
-                              IsTextSearchEnabled="True" 
-                              Height="35" 
-                              Width="393" 
-                              VerticalAlignment="top"
-                              >
-                        <ComboBox.ItemTemplate>
-                            <DataTemplate>
-                                <StackPanel>
-                                    <TextBlock Text="{Binding}" Background="#00185E8D"/>
-                                    <Rectangle Fill="{Binding}"/>
-                                </StackPanel>
-                            </DataTemplate>
-                        </ComboBox.ItemTemplate>
-                    </ComboBox>
-                    <Button Height="33" Click="Button_Click_1">Copy</Button>
-                    <Button Height="33" Click="Button_Click_2">Paste</Button>
-                    <Button Height="33" Click="Button_Click_3">Cut</Button>
-                    <Button Height="33" Click="Button_Click_4">Delete</Button>
-                    <Button Height="33" Click="Button_Click_5">Unzip</Button>
-                    <Button Height="33" Click="Button_Click_6">To Zip</Button>
-                    <!-- Display table for second window-->
-                    <DataGrid x:Name="DataGridOfFilesSecondWindow"
-                              IsReadOnly="True"
-                              AutoGenerateColumns="True"
-                              SelectionUnit="CellOrRowHeader"
-                              VerticalAlignment="Top" Height="275" SelectionChanged="DataGridOfFilesSecondWindow_SelectionChanged"
-                              >
-                        <DataGrid.FrozenColumnCount>1</DataGrid.FrozenColumnCount>
-                    </DataGrid>
-                    <!-- Display search button for second window-->
-                    <ComboBox SelectionChanged="Button_Click" 
-                              x:Name="SelectedFileSecondWindow" 
-                              IsEditable="True" 
-                              IsTextSearchEnabled="True" 
-                              Height="35" 
-                              Width="393" 
-                              VerticalAlignment="top"
-                              >
-                        <ComboBox.ItemTemplate>
-                            <DataTemplate>
-                                <StackPanel>
-                                    <TextBlock Text="{Binding}" Background="#00185E8D"/>
-                                    <Rectangle Fill="{Binding}"/>
-                                </StackPanel>
-                            </DataTemplate>
-                        </ComboBox.ItemTemplate>
-                    </ComboBox>
-                </WrapPanel>
-            </TabItem>
-            <!-- *** Putting start path for search. *** -->
-            <TabItem Header="Search options">
-                <Grid>
-                    <Grid.ColumnDefinitions>
-                        <ColumnDefinition Width="11*"/>
-                        <ColumnDefinition Width="166*"/>
-                        <ColumnDefinition Width="217*"/>
-                    </Grid.ColumnDefinitions>
-                    <TextBlock Margin="100.4,20,99,19.2" Grid.Column="1" 
-                               Text="Search in directory :" Grid.ColumnSpan="2" />
-                    <TextBlock Margin="100.4,40,99,39.2" 
-                               Grid.Column="1" Grid.ColumnSpan="2" >Path :</TextBlock>
-                    <TextBox x:Name="NewPath" 
-                             Margin="130.4,40,324,322.2" 
-                             Grid.Column="1" Grid.ColumnSpan="2" />
-                    <Button Margin="28.4,119,181.8,243.2" 
-                            RenderTransformOrigin="2.961,1.4" 
-                            Click="Button_Click" 
-                            Grid.Column="1">Refresh</Button>
-                    <!---->
-                        <TextBlock Grid.Column="2" 
-                            Text="(Second Window) Search in directory :" Margin="77.4,119,-1,-0.8" />
-                        <TextBlock 
-                            Grid.Column="2" Text="Path :" Margin="37.4,147,82,-40.8" />
-                        <TextBox x:Name="NewPathSecondWindow" 
-                            Grid.Column="2" Margin="77.4,135,34,213.2" />
-                        <Button 
-                            Click="Button_ClickWindowTwo" 
-                            Grid.Column="2" Margin="268.4,183,34,154.2">Refresh second window</Button>
-                </Grid>
-            </TabItem>
-        </TabControl>
-
-
+    <Grid Margin="0,0,-0.4,-30">
+        <local:WindowControl/>
+        <local:WindowControl Margin="383,2,-385,0"/>
     </Grid>
 </Window>
